+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 1.2.3
 
-  * 
+  * Fix bug determining if a badge has been earnt (-1 sections required = all required)
 
 ## Version 1.2.2